Simplify code and support JCad model
